+++ client/src/components/blog/BlogSummary.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import { Container } from 'semantic-ui-react'
-import ItemParagraph from './ItemParagraph';
+import ParagraphContainer from './ParagraphContainer';
 
 export default class BlogSummary extends Component {
   render() {
     return (
       <Container className="blog-item-body">
-        <ItemParagraph summary={ this.props.blog.summary } />
+        <ParagraphContainer summary={ this.props.blog.summary } />
       </Container>
     );
   }
