+++ client/src/components/blog/index.js
@@ -1,5 +1,13 @@
 import BlogContainer from './BlogContainer';
+import BlogMediaContents from './BlogMediaContents';
+import BlogSummary from './BlogSummary';
+import ParagraphContainer from './ParagraphContainer';
+import MediaItem from './MediaItem';
 
 export {
   BlogContainer,
+  BlogMediaContents,
+  BlogSummary,
+  ParagraphContainer,
+  MediaItem,
 };
